Create guess_game.py
This is a simple number guessing game developed as part of my Python practice.  
The user is prompted to guess a randomly generated number within a limited number of attempts.  
The program provides feedback on each guess, helping the user reach the correct number or informing them when attempts are exhausted.